requirements.txt
allure-pytest==2.8.17
allure-python-commons==2.8.17
apipkg==1.5
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
concurrent-log-handler==0.9.9
ConfigArgParse==1.2.3
coverage==5.2.1
ddt==1.4.1
elasticsearch==7.9.1
execnet==1.7.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.9.0
geventhttpclient==1.4.4
greenlet==0.4.17
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==1.4.6
locust==1.3.1
lxml==4.5.2
MarkupSafe==1.1.1
more-itertools==8.4.0
msgpack==1.0.0
nb-log==3.4
packaging==20.4
pika==1.1.0
pluggy==0.13.1
psutil==5.7.2
py==1.9.0
pycparser==2.20
pycryptodome==3.9.8
pymongo==3.5.1
PyMySQL==0.10.1
pyparsing==2.4.7
pypiwin32==223
pytest==6.0.1
pytest-cov==2.10.1
pytest-forked==1.3.0
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-parallel==0.0.10
pytest-rerunfailures==9.1
pytest-sugar==0.9.4
pytest-xdist==2.1.0
PyTestReport==0.2.1
pywin32==228
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
selenium==3.141.0
six==1.15.0
tblib==1.7.0
termcolor==1.1.0
tinydb==4.1.1
toml==0.10.1
tomorrow3==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.1.2
